pyproject.toml
[project]
name = "solana-dexpool-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anchorpy>=0.21.0",
    "boto3>=1.38.6",
    "python-dotenv>=1.1.0",
    "solana>=0.36.6",
    "solders>=0.26.0",
    "whirlpool-essentials>=0.1.0",
]
